cvi_mpi: weekly rls 2026.06.30
ad086a, Merge "[feat](vi):gc4663 support" into v4.2.0
e9ab42, [feat](vi):gc4663 support
942957, [feat](vi): os02h10 bring up

+++ modules/isp/common/raw_dump/src/remote.c
@@ -368,7 +368,7 @@ int send_file(const char *path)
 
 	snprintf(cmd_buffer, CMD_LENGTH, "%s %ld",
 		"cvitek file",
-		statbuf.st_size + FILE_NAME_MAX_LENGTH);
+		(long int)(statbuf.st_size + FILE_NAME_MAX_LENGTH));
 
 	LOGOUT("send cmd: %s\n", cmd_buffer);
